requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
